Added TelegramBot class;

+++ Telegram/Telegram.BotApi/Internal/BindingModels/Update.cs
@@ -8,7 +8,7 @@ namespace Telegram.BotApi.Internal.BindingModels
         /// The update‘s unique identifier. Update identifiers start from a certain positive number and increase sequentially. This ID becomes especially handy if you’re using Webhooks, since it allows you to ignore repeated updates or to restore the correct update sequence, should they get out of order.
         /// </summary>
         [JsonProperty(PropertyName = "update_id")]
-        public int UpdateId { get; set; }
+        public long UpdateId { get; set; }
 
         /// <summary>
         /// Optional. New incoming message of any kind — text, photo, sticker, etc.
